Refactored database and server related code
sherie.klappenbach@medieinstitutet.se ida.bennetsen@medieinstitutet.se cristina.barbagallo@medieinstitutet.se
